+++ AsyncTCPTests/Utilities/Client.m
@@ -44,7 +44,7 @@
     }
     return connect(_descriptor, (struct sockaddr *) &address, sizeof(address));
 }
--(int)send: (char *) data length: (size_t) length {
+-(ssize_t)send: (char *) data length: (size_t) length {
     return send(_descriptor, data, length, 0);
 }
 -(int)close {
